Add abstract news block class
Coding standard - DI for \Drupal calls
