+++ modules/localgov_newsroom/src/Plugin/Block/FeaturedNewsBlock.php
@@ -2,12 +2,6 @@
 
 namespace Drupal\localgov_newsroom\Plugin\Block;
 
-use Drupal\Core\Block\BlockBase;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Drupal\Core\Routing\CurrentRouteMatch;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
 /**
  * Class FeaturedNewsBlock.
  *
@@ -18,67 +12,12 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   admin_label = @Translation("Featured news articles"),
  * )
  */
-class FeaturedNewsBlock extends BlockBase implements ContainerFactoryPluginInterface {
+class FeaturedNewsBlock extends NewsAbstractBlockBase {
 
   /**
    * Maximum number of featured articles.
    */
   const MAX_FEATURED_ARTICLES = 3;
-
-  /**
-   * The entity type manager service.
-   *
-   * @var \Drupal\Core\Entity\EntityTypeManager
-   */
-  protected $entityTypeManager;
-
-  /**
-   * The newsroom node.
-   *
-   * @var \Drupal\node\Entity\Node
-   */
-  protected $newsroom;
-
-  /**
-   * Route match service.
-   *
-   * @var \Drupal\Core\Routing\CurrentRouteMatch
-   */
-  private $routeMatch;
-
-  /**
-   * FeaturedNewsBlock constructor.
-   *
-   * @param array $configuration
-   *   The plugin configuration.
-   * @param string $plugin_id
-   *   The plugin_id for the plugin instance.
-   * @param mixed $plugin_definition
-   *   The plugin implementation definition.
-   * @param \Drupal\Core\Routing\CurrentRouteMatch $route_match
-   *   The route match service.
-   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
-   *   The entity type manager service.
-   */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, CurrentRouteMatch $route_match, EntityTypeManagerInterface $entity_type_manager) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition);
-    $this->routeMatch = $route_match;
-    $this->entityTypeManager = $entity_type_manager;
-    $this->newsroom = $this->routeMatch->getParameter('node');
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->get('current_route_match'),
-      $container->get('entity_type.manager')
-    );
-  }
 
   /**
    * {@inheritdoc}
@@ -126,7 +65,7 @@ class FeaturedNewsBlock extends BlockBase implements ContainerFactoryPluginInter
    * Get the nids for any featured news articles in this newsroom.
    */
   private function getFeaturedNews() {
-    return (array_column($this->newsroom->localgov_newsroom_featured->getValue(), 'target_id'));
+    return (array_column($this->newsroom->loadNewsroom()->get('localgov_newsroom_featured')->getValue(), 'target_id'));
   }
 
   /**
