+++ modules/localgov_newsroom/src/Plugin/Block/NewsAbstractBlockBase.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace Drupal\localgov_newsroom\Plugin\Block;
+
+use Drupal\Core\Access\AccessResult;
+use Drupal\Core\Block\BlockBase;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Drupal\Core\Routing\CurrentRouteMatch;
+use Drupal\Core\Session\AccountInterface;
+use Drupal\localgov_newsroom\Newsroom;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+/**
+ * News contents block.
+ *
+ * @package Drupal\localgov_news\Plugin\Block
+ * )
+ */
+abstract class NewsAbstractBlockBase extends BlockBase implements ContainerFactoryPluginInterface {
+
+  /**
+   * Initialise new content block instance.
+   *
+   * @param array $configuration
+   *   The plugin configuration.
+   * @param string $plugin_id
+   *   The plugin_id for the plugin instance.
+   * @param mixed $plugin_definition
+   *   The plugin implementation definition.
+   * @param \Drupal\Core\Routing\CurrentRouteMatch $route_match
+   *   The route match service.
+   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
+   *   The entity type manager service.
+   * @param \Drupal\localgov_newsroom\Newsroom $newsroom
+   *   The newsroom.
+   */
+  public function __construct(array $configuration, $plugin_id, $plugin_definition, CurrentRouteMatch $route_match, EntityTypeManagerInterface $entity_type_manager, Newsroom $newsroom) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+
+    $this->routeMatch = $route_match;
+    $this->entityTypeManager = $entity_type_manager;
+    $this->newsroom = $newsroom;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('current_route_match'),
+      $container->get('entity_type.manager'),
+      $container->get('localgov_newsroom.newsroom')
+    );
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function blockAccess(AccountInterface $account) {
+    return AccessResult::allowed();
+  }
+
+}
